Integral refactor of Webapp
I've redone everything in the webapp container to handle the twitter 
part (integrating socket etc..) most of the work has been done in 
another git because. well mostly because of internet access problem not 
making able to download this git... sic. 
This is just barebone ATM but all the code for backend is there.